requirements/base.txt
django
djangorestframework
furl
aiohttp>=3.3.2
